chore: add REUSE.toml (#60)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it
Preparation for contribution of OCM to NeoNephos.

This is just a first step, the copyright notices as well as the package
supplier will have to be adjusted, too. This will be done once we know
to what exactly it should be changed!

#### Which issue(s) this PR fixes
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->