chore: add ocm-controller exception (#65)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it

ocm-controller's release process is a remnant of the old days. It has
non-linear commit history by creating merge commits during the release
process. This is sub-optimal, but since we are moving away from it (from
ocm-controller), to fix it is time and effort that we do not want to
spend at this time.

#### Which issue(s) this PR fixes
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->

Signed-off-by: Gergely Brautigam <182850+Skarlso@users.noreply.github.com>