chore: update issue templates (#84)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it
We noticed that we need a lot of roundtrips on bug reports with our
users to get all the information we need to identify and reproduce their
bugs. We hope to be able to reduce those roundtrips by providing a more
verbose description of what we expect in our issue template.

Besides, we remove the issue templates merely used for our project
management from these issue templates (that are propagated to all our
repositories). Instead, we will add them only to `ocm-project`.

#### Which issue(s) this PR is related to
<!--
Usage: `Related to #<issue number>`, or `Related to (paste link of
issue)`.
-->

---------

Signed-off-by: Fabian Burth <fabian.burth@sap.com>
Co-authored-by: Steffen <5519671+sk31337@users.noreply.github.com>